reworking PySD class as a Model object
Stuck on how to pass simulation time to macros.
