+++ pysd/__init__.py
@@ -1,4 +1,4 @@
-from .pysd import read_vensim, load, PySD
+from .pysd import read_vensim, load
 from . import functions
 from . import utils
 from ._version import __version__
